+++ pkg/commander/command.go
@@ -1,4 +1,4 @@
-package cmd
+package commander
 
 import (
 	"fmt"
