+++ pkg/commander/command_test.go
@@ -1,4 +1,4 @@
-package cmd
+package commander
 
 import (
 	"testing"
